# libs/api-tosspaper/openapi.yaml
openapi: 3.1.0
servers:
  - url: http://localhost:8080
    description: Development server
  - url: http://localhost:8080
    description: Local development server
info:
  title: Purchase Order System API
  version: v1
  description: REST API for creating resources in the Purchase Order System.
  contact:
    name: Acme Corp
    url: https://localhost:8080
components:
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "Acme Analytics Platform"
        message:
          type: string
          example: "An error occurred while processing your request"

    # Sender Approval Request
...
    ReviewExtractionRequest:
      type: object
      required:
        - approved
      properties:
        approved:
          type: boolean
          description: |
            - true: Approve the match and create document with provided documentData
            - false: Reject the match (no document created)
          example: true
        notes:
          type: string
          description: Optional notes explaining the approval or rejection decision
          example: "Verified PO number matches invoice. Proceeding with processing."
        documentData:
          type: object
          nullable: true
          description: |
            Document data matching extraction.json schema structure.
            Required when approved=true. Contains the extraction result data (potentially modified by user)
            that will be used to create the Invoice/DeliverySlip/DeliveryNote record.
            Should match the structure from extraction.json schema.
          additionalProperties: true

    # Company
...
    MatchType:
      type: string
      enum: [pending, in_progress, direct, ai_match, no_match, manual, no_po_required]
      description: "Match status: 'pending' (not yet matched), 'in_progress' (AI actively searching), 'direct' (matched by PO number in document), 'ai_match' (AI determined match), 'no_match' (no PO found), 'manual' (manually linked by user), 'no_po_required' (user confirmed no PO needed)"
      example: "pending"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: "Current page number (1-indexed)"
          example: 2
        pageSize:
          type: integer
          description: "Number of items per page"
          example: 20
        totalPages:
          type: integer
          description: "Total number of pages"
          example: 5
        totalItems:
          type: integer
          description: "Total number of items across all pages"
          example: 95
        cursor:
          type: string
          nullable: true
          description: "Cursor for next page (URL-safe base64 encoded created_at|id, generated by backend)"
          example: "MjAyNS0xMS0wMlQyMDozMDoyNi4wNzI1NjMrMDB8MDFLOUM0NzZSVkJTU1NSMllFMTM1VzRERFRX"
...
    CandidateComparison:
      type: object
      description: |
        Contains a detailed field-by-field comparison between a candidate Purchase Order and the extracted document.
        This is the primary data structure AI agents use to understand match quality.
      required:
        - poId
        - poNumber
        - projectId
        - comparison
      properties:
        poId:
          type: string
          description: The unique identifier of the candidate Purchase Order being compared.
          example: "01HQFK8X4J2N3P5R7T9V1W3Y5Z"
        poNumber:
          type: string
          description: The display ID (PO number) of the candidate Purchase Order.
          example: "PO-2024-001"
        projectId:
          type: string
          description: The project ID associated with the candidate Purchase Order.
          example: "01HQFK8X4J2N3P5R7T9V1W3Y5Z"
        comparison:
          $ref: '#/components/schemas/DetailedComparison'
...
    DocumentApproval:
      type: object
      description: "Document approval record for tracking approval workflow."
      properties:
        id:
          type: string
          description: "Approval ID (ULID) for cursor pagination."
          example: "01HQFK8X4J2N3P5R7T9V1W3Y5Z"
        assignedId:
          type: string
          description: "Extraction task identifier (FK to extraction_task.assigned_id)."
          example: "mg-1762125559995-44b16333"
        companyId:
          type: integer
          format: int64
          description: "Company identifier (FK to companies.id)."
        fromEmail:
          type: string
          format: email
          description: "Sender email address."
          example: "vendor@supplier.com"
        documentType:
          type: string
          description: "Document classification (invoice, delivery_slip, etc.)."
          example: "invoice"
        projectId:
          type: string
          nullable: true
          description: "Project identifier associated with the purchase order."
        approvedAt:
          type: string
          format: date-time
          nullable: true
          description: "Timestamp when the document was approved."
        rejectedAt:
          type: string
          format: date-time
          nullable: true
          description: "Timestamp when the document was rejected."
        reviewedBy:
          type: string
          nullable: true
          description: "User identifier who approved/rejected the document."
        reviewNotes:
          type: string
          nullable: true
          description: "Optional notes explaining the approval/rejection decision."
        documentSummary:
          type: string
          nullable: true
          description: "Description of what the document is about."
        poNumber:
          type: string
          nullable: true
          description: "Purchase order number extracted from document"
        documentNumber:
          type: string
          nullable: true
          description: "Document number extracted from document"
        storageKey:
          type: string
          nullable: true
          description: "S3/object storage key for the document file."
        createdAt:
          type: string
          format: date-time
          description: "Timestamp when the approval record was created."
    DocumentApprovalList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DocumentApproval'
        pagination:
          $ref: '#/components/schemas/Pagination'
    DocumentApprovalDetail:
      type: object
      description: "Document approval detail with only the fields used in the UI"
      required:
        - approvalId
        - assignedId
        - companyId
        - documentType
      properties:
        approvalId:
          type: string
          description: "Document approval ID"
        assignedId:
          type: string
          description: "Extraction task ID (assigned_id)"
        companyId:
          type: integer
          format: int64
          description: "Company ID"
        documentType:
          type: string
          description: "Document type (invoice, delivery_slip)"
        externalDocumentNumber:
          type: string
          nullable: true
          description: "External document number (invoice number, delivery ticket number, etc.)"
        poNumber:
          type: string
          nullable: true
          description: "Purchase order number extracted from document"
        fromEmail:
          type: string
          format: email
          description: "Sender email address"
        createdAt:
          type: string
          format: date-time
          description: "When the approval was created"
        projectId:
          type: string
          nullable: true
          description: "Project ID from matched PO"
        approvedAt:
          type: string
          format: date-time
          nullable: true
          description: "When the document was approved"
        rejectedAt:
          type: string
          format: date-time
          nullable: true
          description: "When the document was rejected"
        reviewedBy:
          type: string
          nullable: true
          description: "User ID who reviewed the document"
        reviewNotes:
          type: string
          nullable: true
          description: "Review notes from approver/rejecter"
        documentSummary:
          type: string
          nullable: true
          description: "AI-generated summary of the document"
        storageKey:
          type: string
          nullable: true
          description: "S3 storage key for the document file"
    ExtractionResultResponse:
      type: object
      description: "Raw extraction results from AI processing"
      required:
        - assignedId
        - fromEmail
        - toEmail
        - documentType
        - extractionResult
      properties:
        storageUrl:
          type: string
          description: "storage url"
        assignedId:
          type: string
          description: "Extraction task ID (assigned_id)"
        fromEmail:
          type: string
          format: email
          description: "Sender email address"
        toEmail:
          type: string
          format: email
          description: "Recipient email address"
        documentType:
          type: string
          description: "Document type (invoice, delivery_slip, delivery_note)"
        extractionResult:
          type: string
          description: "Raw extraction results from provider API (extract_task_results JSONB)"
        matchType:
          $ref: '#/components/schemas/MatchType'
          nullable: true
        poNumber:
          type: string
        projectId:
          type: string
        poId:
          type: string
        matchReport:
          type: array
          description: "Array of detailed comparisons for all candidate POs evaluated."
          items:
            $ref: '#/components/schemas/CandidateComparison'
...
    LinkPoRequest:
      type: object
      description: |
        Request payload for manually linking a document to a specific Purchase Order.
        Used when automatic matching needs to be overridden or when user manually selects a PO.
      required:
        - poNumber
      properties:
        poNumber:
          type: string
          description: |
            The unique identifier of the Purchase Order to link to the document.
            This PO must:
            - Already exist in the system (created via POST /v1/projects/{projectId}/purchase-orders)
            - Be ingested into the vector store (happens automatically after creation)
            - Belong to the same company as the document (validated by to_email matching)
            
            If the PO is not found in the vector store, the system will create a NO_MATCH result.
          example: "01HQFK8X4J2N3P5R7T9V1W3Y5Z"
...
  responses:
    BadRequest:
      description: Bad Request - The request could not be understood by the server due to malformed syntax.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - Authentication credentials are required and have failed or have not yet been provided.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not Found - The requested resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal Server Error - A generic error message, given when an unexpected condition was encountered.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - jwtAuth: []
paths:
...
  /v1/document-matches/{assignedId}/link-po:
    post:
      operationId: linkPurchaseOrder
      tags:
        - "Document Matching"
      summary: "Manually link a document to a specific Purchase Order"
      description: |
        Manually override automatic matching by linking an extracted document to a specific PO.
        
        **Use cases:**
        - User selects a different PO from the candidate list
        - Automatic matching failed but user knows the correct PO
        - User wants to re-link a document to a different PO
        
        **Process:**
        1. Validates the specified PO exists and belongs to the company
        2. Performs detailed field-by-field comparison between document and PO
        3. Creates/updates a DocumentMatch record with match_type=DIRECT
        4. Returns 202 Accepted immediately (processing happens asynchronously)

        **Important notes:**
        - The PO must already be in the vector store (i.e., created via the system)
        - If PO is not found in vector store, returns NO_MATCH
        - Creates an upsert - if match already exists for this document, it will be replaced
        
        **AI Integration Tips:**
        - Use this endpoint when user manually selects a PO from suggestions
        - Check the resulting match_report to show comparison details to user
        - Poll the extraction endpoint after ~2-5 seconds to get results
      parameters:
        - name: X-Context-Id
          in: header
          required: true
          schema:
            type: string
          description: "Context header containing the authenticated company identifier."

        - name: assignedId
          in: path
          required: true
          description: |
            The unique ID of the extraction task (assigned_id).
            This is the document you want to link to a PO.
          schema:
            type: string
          example: "01HQFK8X4J2N3P5R7T9V1W3Y5Z"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkPoRequest'
      responses:
        "202":
          description: "Accepted. Request queued for async processing."
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalServerError'
  /v1/document-matches/{assignedId}/rematch:
    post:
      operationId: rematch
      tags:
        - "Document Matching"
      summary: "Re-run automatic PO matching for a document"
      description: |
        Trigger a fresh automatic matching process for an already-extracted document.
        
        **Use cases:**
        - New POs were added to the system after the document was processed
        - Previous automatic match was unsatisfactory
        - User rejected the current match and wants to see other options
        - Initial matching failed (NO_MATCH) but POs are now available
        
        **Process:**
        1. Re-runs the complete matching workflow:
           - Direct match by PO number
           - Vector similarity search for candidates
           - AI evaluation of all viable candidates
        2. Creates/updates DocumentMatch record with new results
        3. Returns 202 Accepted immediately (processing happens asynchronously)
        
        **Important notes:**
        - This is a full re-match, not just a retry of the previous attempt
        - May find different POs if new ones were added to the system
        - Will replace any existing match for this document (upsert)
        - May take 5-15 seconds depending on number of candidate POs
        
        **AI Integration Tips:**
        - Use this when user is unhappy with current match suggestions
        - Check if new POs were recently created before calling this
        - Show "Re-analyzing..." status to user while processing
        - Poll the extraction endpoint after ~5-10 seconds for results
        - Compare new match_report with previous to show what changed
      parameters:
        - name: assignedId
          in: path
          required: true
          description: |
            The unique ID of the extraction task (assigned_id).
            This is the document you want to re-match.
          schema:
            type: string
          example: "01HQFK8X4J2N3P5R7T9V1W3Y5Z"
      responses:
        "202":
          description: "Accepted. Re-match request queued for async processing."
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalServerError'
  /v1/document-approvals:
    get:
      operationId: listDocumentApprovals
      tags:
        - "Document Approvals"
      summary: "List Document Approvals"
      description: |
        Retrieves a paginated view of document approvals with cursor-based pagination.
        These represent documents awaiting approval or already approved/rejected.

        Results are automatically scoped to the company ID from the `X-Context-Id` header.
        Results are ordered by:
        - Approved documents: ordered by `approved_at` DESC
        - Rejected documents: ordered by `rejected_at` DESC  
        - Pending documents: ordered by `created_at` DESC
      parameters:
        - name: X-Context-Id
          in: header
          required: true
          schema:
            type: string
          description: "Context header containing the authenticated company identifier."
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
          description: "Number of items per page."
        - name: cursor
          in: query
          required: false
          schema:
            type: string
          description: "Cursor for pagination (URL-safe base64 encoded timestamp|id from previous page)"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [approved, rejected, pending]
          description: "Filter by approval status."
        - name: documentType
          in: query
          required: false
          schema:
            type: string
          description: "Filter by document type (e.g., invoice, delivery_slip)."
        - name: fromEmail
          in: query
          required: false
          schema:
            type: string
            format: email
          description: "Filter by sender email address."
        - name: createdDateFrom
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: "Filter by approval creation timestamp (inclusive, UTC)."
        - name: createdDateTo
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: "Filter by approval creation timestamp (inclusive, UTC)."
        - name: projectId
          in: query
          required: false
          schema:
            type: string
          description: "Filter by purchase order project identifier."
      responses:
        "200":
          description: Successfully retrieved document approvals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentApprovalList'
        "401":
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/InternalServerError'
  /v1/document-approvals/{approvalId}:
    get:
      operationId: getDocumentApprovalDetail
      tags:
        - "Document Approvals"
      summary: "Get Document Approval Detail"
      description: |
        Retrieves complete document approval detail for review page.
        Includes extraction metadata, document content, match information, and approval status.

        This endpoint combines data from:
        - document_approvals (approval status, review info)
        - extraction_task (email metadata, extracted content)
        - invoices/delivery_slips (match information, AI comparison report)
      parameters:
        - name: X-Context-Id
          in: header
          required: true
          schema:
            type: string
          description: "Context header containing the authenticated company identifier."
        - name: approvalId
          in: path
          required: true
          description: Document approval ID (from document_approvals.id)
          schema:
            type: string
      responses:
        "200":
          description: Successfully retrieved document approval detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentApprovalDetail'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalServerError'
  /v1/document-approvals/{approvalId}/extraction:
    get:
      operationId: getDocumentApprovalExtraction
      tags:
        - "Document Approvals"
      summary: "Get raw extraction results for document approval"
      description: |
        Retrieves the raw extraction results from the AI extraction process.
        This endpoint exposes the unformatted extraction data that the user can review and modify
        before approving the document.

        Returns:
        - Raw extraction results (extract_task_results JSONB from extraction_task table)
        - Email metadata (from/to addresses)
        - Document type and assigned ID

        This data is used by the frontend to show the user what was extracted and allow them
        to make modifications before submitting the approval with the corrected data.
      parameters:
        - name: X-Context-Id
          in: header
          required: true
          schema:
            type: string
          description: "Context header containing the authenticated company identifier."
        - name: approvalId
          in: path
          required: true
          description: Document approval ID (from document_approvals.id)
          schema:
            type: string
      responses:
        "200":
          description: Successfully retrieved extraction results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractionResultResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/InternalServerError'
  /v1/document-approvals/{approvalId}/extractions:
    post:
      operationId: reviewExtraction
      tags:
        - "Document Approvals"
      summary: "Review and approve or reject an extraction"
      description: |
        Review and approve or reject a document approval. When approved, the matched 
        purchase order status will be automatically updated to IN_PROGRESS (if currently PENDING),
        and invoice/delivery slip records will be created asynchronously.
        
        **Use cases:**
        - User confirms the document match is correct and wants to proceed
        - User rejects the document due to incorrect PO linking or other issues
        - User provides notes explaining their decision
        
        **Behavior:**
        - Approved documents: Sets approved_at timestamp and publishes event for async processing
          - PO status transitions to IN_PROGRESS (if currently PENDING)
          - Invoice/delivery slip records created asynchronously via Redis stream
          - Acceptance email sent to document sender
        - Rejected documents: Sets rejected_at timestamp and sends rejection email
          - PO status remains unchanged
          - No records created
        - Notes are optional but recommended for rejected documents
        
        **Important notes:**
        - The path parameter is actually an approval ID (ULID) from document_approvals table
        - Rejected documents can be approved later (rejected → approved transition allowed)
        - Approved documents cannot be rejected (approved → rejected transition blocked)
        - PO status transition only occurs if PO is currently PENDING
        - If PO is already IN_PROGRESS, COMPLETED, or CANCELLED, no status change occurs
      parameters:
        - name: X-Context-Id
          in: header
          required: true
          description: Company ID
          schema:
            type: string
        - name: approvalId
          in: path
          required: true
          description: Approval ID (ULID) from the document_approvals table
          schema:
            type: string
            example: "01HQFK8X4J2N3P5R7T9V1W3Y5Z"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewExtractionRequest'
      responses:
        "204":
          description: Review processed successfully
        "400":
          description: Bad request - cannot reject an already approved document
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          description: Document approval not found or access denied (wrong company)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        default:
          $ref: '#/components/responses/InternalServerError'
...